move get_auth_user to auth.py
